color-difference-feature-start-rael-opencv/requirements.txt
colour-science>=0.4.2
gradio==5.35.0
gradio-client==1.10.4
matplotlib>=3.7.0
numpy>=2.2.0
opencv-contrib-python>=4.8.0
pillow>=11.0.0
PyYAML>=6.0.0
